Add patchy compatibility with symfony/scheduler >= 6.4
Updated the TaskInfo.php
to introduce compatibility with symfony/scheduler version 6.4 and up.
The code checks if the method 'getMessage' exists for task objects,
to support both newer and older versions of the scheduler.
Also,
cleared out the comments in schedule.html.twig
to reveal details about message type,
description, and redispatch transports.